Update: Move section to pages/home
